refactor(tests): split analytics-coverage.test.ts into 4 focused files
Splits the 2289-line monolith into:
- analytics-coverage-init: initAnalytics, onFirstInteraction, loadGtagScript, loadUmamiScript, sendToUmami, send() gating
- analytics-coverage-send: sendViaProxy, sendViaGtag, flushPendingEvents, engagement tracking integration
- analytics-coverage-tracking: hashUserId, tryChunkReloadRecovery, global error tracking filters, captureUtmParams, setAnalyticsOptOut, updateAnalyticsIds, setAnalyticsUserProperties, emitDemoModeToggled, emitSessionContext, emitDeveloperSession
- analytics-coverage-emit: emitAgentProvidersDetected, NPS, Orbit/GroundControl, emitBlogPostClicked, engagement session tracking, engagement time, markGtagDecided, wasAlreadyReported

Fixes #8624
