--- web/src/lib/__tests__/analytics-coverage.test.ts
@@ -1,2304 +1,0 @@
-/**
- * Coverage tests for analytics.ts — targets ~229 uncovered lines.
- *
- * Uses vi.resetModules() + dynamic import to get fresh module state for each
- * test group, allowing us to exercise initialization, gtag loading, engagement
- * tracking, error handlers, and proxy/gtag send paths with clean state.
- */
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
-
-/**
- * Returns true if the element's src URL has the given hostname — uses
- * new URL() instead of includes() to prevent CodeQL
- * js/incomplete-url-substring-sanitization false positives (#9119).
- */
-function srcHasHostname(el: Element, hostname: string): boolean {
-  const src = (el as HTMLScriptElement).src
-  if (!src) return false
-  try {
-    return new URL(src).hostname.toLowerCase() === hostname.toLowerCase()
-  } catch {
-    return false
-  }
-}
-
-// ── Shared mock setup ──────────────────────────────────────────────
-
-vi.mock('../constants', async (importOriginal) => {
-  const actual = (await importOriginal()) as Record<string, unknown>
-  return {
-    ...actual,
-    STORAGE_KEY_ANALYTICS_OPT_OUT: 'kc-analytics-opt-out',
-    STORAGE_KEY_ANONYMOUS_USER_ID: 'kc-anonymous-user-id',
-  }
-})
-
-vi.mock('../chunkErrors', () => ({
-  CHUNK_RELOAD_TS_KEY: 'ksc-chunk-reload-ts',
-  isChunkLoadMessage: (msg: string) =>
-    msg.includes('Failed to fetch dynamically imported module') ||
-    msg.includes('Loading chunk') ||
-    msg.includes('Importing a module script failed'),
-}))
-
-let mockIsDemoMode = false
-let mockIsNetlifyDeployment = false
-
-vi.mock('../demoMode', () => ({
-  isDemoMode: () => mockIsDemoMode,
-  get isNetlifyDeployment() {
-    return mockIsNetlifyDeployment
-  },
-}))
-
-// ── Helper: fresh import ──────────────────────────────────────────
-
-type AnalyticsModule = typeof import('../analytics')
-
-async function freshImport(): Promise<AnalyticsModule> {
-  vi.resetModules()
-  return (await import('../analytics')) as AnalyticsModule
-}
-
-// ── Setup / Teardown ──────────────────────────────────────────────
-
-beforeEach(() => {
-  localStorage.clear()
-  sessionStorage.clear()
-  mockIsDemoMode = false
-  mockIsNetlifyDeployment = false
-  vi.useFakeTimers({ shouldAdvanceTime: false })
-
-  // Provide baseline DOM APIs that analytics.ts expects
-  vi.stubGlobal('navigator', {
-    ...navigator,
-    webdriver: false,
-    userAgent: 'Mozilla/5.0 Chrome/120.0',
-    plugins: { length: 2 },
-    languages: ['en-US'],
-    language: 'en-US',
-    sendBeacon: vi.fn(() => true),
-  })
-
-  // Clean up any gtag globals from prior tests
-  delete (window as Record<string, unknown>).dataLayer
-  delete (window as Record<string, unknown>).gtag
-  delete (window as Record<string, unknown>).google_tag_manager
-  delete (window as Record<string, unknown>).umami
-})
-
-afterEach(() => {
-  vi.useRealTimers()
-  vi.restoreAllMocks()
-  vi.unstubAllGlobals()
-})
-
-// ============================================================================
-// initAnalytics — bot detection, initialization gating
-// ============================================================================
-
-describe('initAnalytics with fresh module state', () => {
-  it('sets initialized=true and registers interaction listeners', async () => {
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    // Should have registered interaction gate events
-    const interactionCalls = addSpy.mock.calls.filter(
-      ([evt]) =>
-        evt === 'mousedown' ||
-        evt === 'keydown' ||
-        evt === 'scroll' ||
-        evt === 'touchstart' ||
-        evt === 'click',
-    )
-    expect(interactionCalls.length).toBeGreaterThanOrEqual(5)
-  })
-
-  it('skips initialization in WebDriver environment', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: true,
-      userAgent: 'Mozilla/5.0 Chrome/120.0',
-      plugins: { length: 2 },
-      languages: ['en-US'],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    // Should NOT register interaction listeners if automated
-    const interactionCalls = addSpy.mock.calls.filter(
-      ([evt]) => evt === 'mousedown' || evt === 'click',
-    )
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('skips initialization in HeadlessChrome environment', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: false,
-      userAgent: 'Mozilla/5.0 HeadlessChrome/120.0',
-      plugins: { length: 2 },
-      languages: ['en-US'],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('skips initialization in PhantomJS environment', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: false,
-      userAgent: 'Mozilla/5.0 PhantomJS/2.1',
-      plugins: { length: 2 },
-      languages: ['en-US'],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('skips initialization when no plugins (non-Firefox)', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: false,
-      userAgent: 'Mozilla/5.0 Chrome/120.0',
-      plugins: { length: 0 },
-      languages: ['en-US'],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('does NOT skip Firefox with no plugins', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: false,
-      userAgent: 'Mozilla/5.0 Firefox/120.0',
-      plugins: { length: 0 },
-      languages: ['en-US'],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    expect(interactionCalls.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('skips initialization when no languages', async () => {
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      webdriver: false,
-      userAgent: 'Mozilla/5.0 Chrome/120.0',
-      plugins: { length: 2 },
-      languages: [],
-      language: 'en-US',
-      sendBeacon: vi.fn(),
-    })
-    const mod = await freshImport()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics()
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('is idempotent — second call is a no-op', async () => {
-    const mod = await freshImport()
-    mod.initAnalytics()
-    const addSpy = vi.spyOn(document, 'addEventListener')
-    mod.initAnalytics() // second call
-    const interactionCalls = addSpy.mock.calls.filter(([evt]) => evt === 'mousedown')
-    // Second call should NOT register additional listeners
-    expect(interactionCalls.length).toBe(0)
-  })
-
-  it('registers beforeunload and global error tracking', async () => {
-    const windowAddSpy = vi.spyOn(window, 'addEventListener')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    const beforeUnloadCalls = windowAddSpy.mock.calls.filter(([evt]) => evt === 'beforeunload')
-    expect(beforeUnloadCalls.length).toBeGreaterThanOrEqual(1)
-    const errorCalls = windowAddSpy.mock.calls.filter(([evt]) => evt === 'error')
-    expect(errorCalls.length).toBeGreaterThanOrEqual(1)
-    const rejectionCalls = windowAddSpy.mock.calls.filter(
-      ([evt]) => evt === 'unhandledrejection',
-    )
-    expect(rejectionCalls.length).toBeGreaterThanOrEqual(1)
-  })
-})
-
-// ============================================================================
-// onFirstInteraction — script loading, event flushing, pending recovery
-// ============================================================================
-
-describe('onFirstInteraction triggers script loading and flushing', () => {
-  it('loads gtag and umami scripts on first mousedown', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-
-    // Simulate user interaction
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Should have appended script elements (gtag + umami)
-    const scriptAppends = appendSpy.mock.calls.filter(
-      ([el]) => el instanceof HTMLScriptElement,
-    )
-    expect(scriptAppends.length).toBeGreaterThanOrEqual(2)
-  })
-
-  it('is idempotent — second interaction does not re-load scripts', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-
-    document.dispatchEvent(new Event('mousedown'))
-    const countAfterFirst = appendSpy.mock.calls.filter(
-      ([el]) => el instanceof HTMLScriptElement,
-    ).length
-
-    document.dispatchEvent(new Event('mousedown'))
-    const countAfterSecond = appendSpy.mock.calls.filter(
-      ([el]) => el instanceof HTMLScriptElement,
-    ).length
-
-    expect(countAfterSecond).toBe(countAfterFirst)
-  })
-
-  it('flushes pending recovery event on first interaction', async () => {
-    // Set up a chunk-reload recovery marker BEFORE module import
-    const reloadTime = Date.now() - 500
-    sessionStorage.setItem('ksc-chunk-reload-ts', String(reloadTime))
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-
-    // Recovery should be detected but not sent yet (user hasn't interacted)
-    // Marker should be cleared from sessionStorage by initAnalytics -> startGlobalErrorTracking -> checkChunkReloadRecovery
-    expect(sessionStorage.getItem('ksc-chunk-reload-ts')).toBeNull()
-
-    // Trigger interaction — should flush the pending recovery event
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', { ...navigator, sendBeacon: beaconSpy, language: 'en-US' })
-    document.dispatchEvent(new Event('mousedown'))
-
-    // The recovery event should have been emitted (either via beacon or queued for gtag)
-    // Since gtag hasn't loaded yet, events are queued
-    // This verifies the code path doesn't throw
-  })
-})
-
-// ============================================================================
-// loadGtagScript — script loading, CDN fallback, timeout
-// ============================================================================
-
-describe('loadGtagScript behavior', () => {
-  it('creates script element with first-party proxy src', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    const scripts = appendSpy.mock.calls
-      .filter(([el]) => el instanceof HTMLScriptElement)
-      .map(([el]) => (el as HTMLScriptElement).src)
-
-    // Should have the gtag proxy script
-    const gtagScript = scripts.find((s) => s.includes('/api/gtag'))
-    expect(gtagScript).toBeTruthy()
-  })
-
-  it('initializes dataLayer and gtag function', async () => {
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    expect(window.dataLayer).toBeDefined()
-    expect(Array.isArray(window.dataLayer)).toBe(true)
-    expect(typeof window.gtag).toBe('function')
-  })
-
-  it('falls back to CDN on script.onerror', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Find the first-party proxy script and trigger its onerror
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    expect(firstScript).toBeTruthy()
-    if (firstScript?.onerror) {
-      ;(firstScript.onerror as () => void)()
-    }
-
-    // Should have appended a CDN fallback script
-    const cdnScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement &&
-          srcHasHostname(el, 'www.googletagmanager.com'),
-      ) as HTMLScriptElement | undefined
-
-    expect(cdnScript).toBeTruthy()
-  })
-
-  it('falls back to CDN when proxy returns HTML (not real gtag)', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    expect(firstScript).toBeTruthy()
-
-    // Simulate onload without google_tag_manager being set (HTML response)
-    // window.google_tag_manager is NOT defined
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-
-    // Advance past GTAG_INIT_CHECK_MS (100ms)
-    vi.advanceTimersByTime(150)
-
-    // Should have appended CDN fallback
-    const cdnScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement &&
-          srcHasHostname(el, 'www.googletagmanager.com'),
-      )
-
-    expect(cdnScript).toBeTruthy()
-  })
-
-  it('marks gtag as available when proxy loads successfully', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Simulate successful gtag.js initialization
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-
-    vi.advanceTimersByTime(150)
-
-    // After gtag decided=true+available, events should go through sendViaGtag
-    // Test by emitting an event — should call window.gtag
-    const gtagFn = vi.fn()
-    window.gtag = gtagFn
-    mod.emitPageView('/test')
-    expect(gtagFn).toHaveBeenCalled()
-  })
-
-  it('CDN fallback onerror marks gtag as unavailable', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Trigger first-party onerror
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onerror) {
-      ;(firstScript.onerror as () => void)()
-    }
-
-    // Now trigger CDN onerror
-    const cdnScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement &&
-          srcHasHostname(el, 'www.googletagmanager.com'),
-      ) as HTMLScriptElement | undefined
-
-    if (cdnScript?.onerror) {
-      ;(cdnScript.onerror as () => void)()
-    }
-
-    // Now events should go via proxy (sendBeacon)
-    mod.emitPageView('/test')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('timeout falls back to proxy when gtag.js takes too long', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Advance past GTAG_LOAD_TIMEOUT_MS (5000ms) without script loading
-    vi.advanceTimersByTime(5100)
-
-    // Events should now go via proxy
-    mod.emitPageView('/test')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// loadUmamiScript — script creation
-// ============================================================================
-
-describe('loadUmamiScript', () => {
-  it('creates umami script with correct attributes', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    const umamiScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/ksc'),
-      ) as HTMLScriptElement | undefined
-
-    expect(umamiScript).toBeTruthy()
-    expect(umamiScript?.defer).toBe(true)
-    expect(umamiScript?.dataset.websiteId).toBeTruthy()
-    expect(umamiScript?.dataset.hostUrl).toBe(window.location.origin)
-  })
-})
-
-// ============================================================================
-// sendToUmami — fire-and-forget
-// ============================================================================
-
-describe('sendToUmami', () => {
-  it('calls umami.track when available', async () => {
-    const trackFn = vi.fn()
-    ;(window as Record<string, unknown>).umami = { track: trackFn }
-
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Force gtag decided so events aren't just queued
-    vi.advanceTimersByTime(5100)
-
-    mod.emitPageView('/umami-test')
-    expect(trackFn).toHaveBeenCalledWith(
-      'page_view',
-      expect.objectContaining({ page_path: '/umami-test' }),
-    )
-  })
-
-  it('does not throw when umami is undefined', async () => {
-    delete (window as Record<string, unknown>).umami
-
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    expect(() => mod.emitPageView('/test')).not.toThrow()
-  })
-
-  it('does not throw when umami.track throws', async () => {
-    ;(window as Record<string, unknown>).umami = {
-      track: () => {
-        throw new Error('Umami failure')
-      },
-    }
-
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    expect(() => mod.emitPageView('/test')).not.toThrow()
-  })
-})
-
-// ============================================================================
-// send() gating — not initialized, opted out, not interacted
-// ============================================================================
-
-describe('send() gating', () => {
-  it('drops events when not initialized', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    // Do NOT call initAnalytics
-    mod.emitPageView('/test')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('drops events when opted out', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    // Clear beacon calls from initialization events (page_view, conversion_step)
-    beaconSpy.mockClear()
-
-    localStorage.setItem('kc-analytics-opt-out', 'true')
-    mod.emitPageView('/test')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('drops events before user interaction', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    // Do NOT dispatch interaction event
-    mod.emitPageView('/test')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('queues events while waiting for gtag decision', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Emit events while gtag is still loading (before timeout)
-    mod.emitCardAdded('pods', 'manual')
-    mod.emitCardRemoved('nodes')
-
-    // Events should NOT have been sent via beacon yet (queued)
-    // The initial page_view + conversion from onFirstInteraction are also queued
-    // but no beacon calls since gtag hasn't decided
-    const beaconCallsBefore = beaconSpy.mock.calls.length
-
-    // Now let gtag timeout to flush queue via proxy
-    vi.advanceTimersByTime(5100)
-
-    // Queue should have been flushed via beacon
-    expect(beaconSpy.mock.calls.length).toBeGreaterThan(beaconCallsBefore)
-  })
-})
-
-// ============================================================================
-// sendViaProxy — parameter encoding, UTM, engagement time
-// ============================================================================
-
-describe('sendViaProxy detailed behavior', () => {
-  async function setupProxyMode() {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    // Force gtag to be unavailable so proxy is used
-    vi.advanceTimersByTime(5100)
-    // Clear beacon calls from initialization events
-    beaconSpy.mockClear()
-    return { mod, beaconSpy }
-  }
-
-  it('encodes event payload as base64 in query string', async () => {
-    const { mod, beaconSpy } = await setupProxyMode()
-    mod.emitCardAdded('pods', 'manual')
-    expect(beaconSpy).toHaveBeenCalledTimes(1)
-    const url = beaconSpy.mock.calls[0][0] as string
-    expect(url).toContain('/api/m?d=')
-    // The d= parameter should be base64 encoded
-    const encoded = decodeURIComponent(url.split('d=')[1])
-    const decoded = atob(encoded)
-    expect(decoded).toContain('en=ksc_card_added')
-    expect(decoded).toContain('ep.card_type=pods')
-    expect(decoded).toContain('ep.source=manual')
-  })
-
-  it('uses epn. prefix for numeric params', async () => {
-    const { mod, beaconSpy } = await setupProxyMode()
-    mod.emitCardPaginationUsed(3, 10, 'pods')
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('epn.page=3')
-    expect(decoded).toContain('epn.total_pages=10')
-    expect(decoded).toContain('ep.card_type=pods')
-  })
-
-  it('includes UTM params when captured', async () => {
-    // Set up URL with UTM params
-    const originalSearch = window.location.search
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        search: '?utm_source=github&utm_medium=social&utm_campaign=launch',
-        href: 'http://localhost/?utm_source=github&utm_medium=social&utm_campaign=launch',
-        hostname: 'localhost',
-        pathname: '/',
-        origin: 'http://localhost',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitPageView('/')
-    const url = beaconSpy.mock.calls[0]?.[0] as string
-    if (url) {
-      const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-      expect(decoded).toContain('cs=github')
-      expect(decoded).toContain('cm=social')
-      expect(decoded).toContain('cn=launch')
-    }
-
-    // Restore
-    Object.defineProperty(window, 'location', {
-      value: { ...window.location, search: originalSearch },
-      writable: true,
-      configurable: true,
-    })
-  })
-
-  it('falls back to fetch when sendBeacon is unavailable', async () => {
-    const fetchSpy = vi.fn(() => Promise.resolve(new Response()))
-    vi.stubGlobal('fetch', fetchSpy)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: undefined,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    mod.emitCardAdded('test', 'manual')
-    expect(fetchSpy).toHaveBeenCalled()
-    const [url, opts] = fetchSpy.mock.calls[fetchSpy.mock.calls.length - 1]
-    expect(url).toContain('/api/m?d=')
-    expect((opts as RequestInit).method).toBe('POST')
-    expect((opts as RequestInit).keepalive).toBe(true)
-  })
-
-  it('includes user ID when set', async () => {
-    const { mod, beaconSpy } = await setupProxyMode()
-    await mod.setAnalyticsUserId('real-user-123')
-    beaconSpy.mockClear()
-
-    mod.emitPageView('/test')
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('uid=')
-  })
-
-  it('sets _fv=1 on very first session', async () => {
-    // Ensure no prior sessions exist
-    localStorage.clear()
-    const { mod, beaconSpy } = await setupProxyMode()
-    beaconSpy.mockClear()
-
-    mod.emitPageView('/test')
-    if (beaconSpy.mock.calls.length > 0) {
-      const url = beaconSpy.mock.calls[0][0] as string
-      const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-      // First visit flag may or may not be set depending on session state from init
-      // This just exercises the code path
-      expect(decoded).toContain('v=2')
-    }
-  })
-
-  it('sets _ss and _nsi on new sessions', async () => {
-    const { mod, beaconSpy } = await setupProxyMode()
-    // Force session to expire
-    localStorage.setItem('_ksc_last', String(Date.now() - 31 * 60 * 1000))
-    beaconSpy.mockClear()
-
-    mod.emitPageView('/test')
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('_ss=1')
-    expect(decoded).toContain('_nsi=1')
-  })
-
-  it('includes user properties in proxy payload', async () => {
-    const { mod, beaconSpy } = await setupProxyMode()
-    mod.setAnalyticsUserProperties({ role: 'admin', team: 'platform' })
-    beaconSpy.mockClear()
-
-    mod.emitPageView('/test')
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('up.role=admin')
-    expect(decoded).toContain('up.team=platform')
-  })
-})
-
-// ============================================================================
-// sendViaGtag — engagement time, user ID
-// ============================================================================
-
-describe('sendViaGtag detailed behavior', () => {
-  async function setupGtagMode() {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Simulate successful gtag load
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-
-    const gtagSpy = vi.fn()
-    window.gtag = gtagSpy
-    return { mod, gtagSpy }
-  }
-
-  it('sends events through window.gtag', async () => {
-    const { mod, gtagSpy } = await setupGtagMode()
-    mod.emitCardAdded('pods', 'manual')
-    expect(gtagSpy).toHaveBeenCalledWith(
-      'event',
-      'ksc_card_added',
-      expect.objectContaining({ card_type: 'pods', source: 'manual' }),
-    )
-  })
-
-  it('includes user_id in gtag events when set', async () => {
-    const { mod, gtagSpy } = await setupGtagMode()
-    await mod.setAnalyticsUserId('user-xyz')
-    gtagSpy.mockClear()
-
-    mod.emitCardAdded('test', 'manual')
-    expect(gtagSpy).toHaveBeenCalledWith(
-      'event',
-      'ksc_card_added',
-      expect.objectContaining({ user_id: expect.any(String) }),
-    )
-  })
-
-  it('does not send when window.gtag is missing', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-
-    // Remove gtag after it was decided as available
-    delete (window as Record<string, unknown>).gtag
-    expect(() => mod.emitCardAdded('test', 'manual')).not.toThrow()
-  })
-})
-
-// ============================================================================
-// flushPendingEvents — queued events flushed via gtag or proxy
-// ============================================================================
-
-describe('flushPendingEvents', () => {
-  it('flushes queued events via gtag when available', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Events are queued while gtag decision pending
-    // The onFirstInteraction already fires page_view + conversion_step
-
-    // Simulate successful gtag load to flush via gtag
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const gtagSpy = vi.fn()
-    window.gtag = gtagSpy
-
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-
-    // Queued events from onFirstInteraction should have been flushed via gtag
-    const eventCalls = gtagSpy.mock.calls.filter(([type]) => type === 'event')
-    expect(eventCalls.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('flushes queued events via proxy when gtag unavailable', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    beaconSpy.mockClear()
-
-    // Let gtag timeout
-    vi.advanceTimersByTime(5100)
-
-    // Queued events from onFirstInteraction should have been flushed via beacon
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// Engagement tracking — markActive, checkEngagement, visibility change
-// ============================================================================
-
-describe('engagement tracking integration', () => {
-  it('starts engagement tracking on first interaction', async () => {
-    const docAddSpy = vi.spyOn(document, 'addEventListener')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Should have registered engagement interaction listeners
-    const engagementCalls = docAddSpy.mock.calls.filter(
-      ([evt]) =>
-        evt === 'visibilitychange' ||
-        (typeof evt === 'string' && ['mousedown', 'keydown', 'scroll', 'touchstart'].includes(evt)),
-    )
-    expect(engagementCalls.length).toBeGreaterThanOrEqual(4)
-  })
-
-  it('starts heartbeat interval for idle detection', async () => {
-    const setIntervalSpy = vi.spyOn(globalThis, 'setInterval')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // setInterval should have been called for heartbeat (5000ms)
-    const heartbeatCalls = setIntervalSpy.mock.calls.filter(
-      ([, ms]) => ms === 5000,
-    )
-    expect(heartbeatCalls.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('visibility hidden flushes engagement and marks inactive', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100) // force proxy mode
-    beaconSpy.mockClear()
-
-    // Advance time so there is some engagement accumulated
-    vi.advanceTimersByTime(2000)
-
-    // Simulate tab hidden
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'hidden',
-      configurable: true,
-    })
-    document.dispatchEvent(new Event('visibilitychange'))
-
-    // Should have emitted user_engagement event
-    const engagementCall = beaconSpy.mock.calls.find(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=user_engagement')
-    })
-    // The engagement event may or may not fire depending on accumulated time
-    // but the code path should not throw
-    expect(true).toBe(true)
-
-    // Restore
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'visible',
-      configurable: true,
-    })
-  })
-
-  it('visibility visible re-marks active', async () => {
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    // Hidden then visible
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'hidden',
-      configurable: true,
-    })
-    document.dispatchEvent(new Event('visibilitychange'))
-
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'visible',
-      configurable: true,
-    })
-    document.dispatchEvent(new Event('visibilitychange'))
-
-    // Should not throw
-    expect(true).toBe(true)
-  })
-
-  it('heartbeat detects idle user after 60s of no interaction', async () => {
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100) // force proxy mode
-
-    // Advance past idle threshold (60s) without any interaction
-    vi.advanceTimersByTime(65000)
-
-    // The heartbeat should have called checkEngagement which sets isUserActive=false
-    // This doesn't throw
-    expect(true).toBe(true)
-  })
-
-  it('stopEngagementTracking clears heartbeat on opt-out', async () => {
-    const clearIntervalSpy = vi.spyOn(globalThis, 'clearInterval')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    mod.setAnalyticsOptOut(true)
-    expect(clearIntervalSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// hashUserId — crypto.subtle and FNV fallback
-// ============================================================================
-
-describe('hashUserId via setAnalyticsUserId', () => {
-  it('uses crypto.subtle when available (default in tests)', async () => {
-    const mod = await freshImport()
-    await mod.setAnalyticsUserId('test-user')
-    // Should not throw — crypto.subtle is available in vitest
-  })
-
-  it('uses FNV fallback when crypto.subtle is undefined', async () => {
-    const originalCrypto = globalThis.crypto
-    // Stub crypto without subtle
-    vi.stubGlobal('crypto', {
-      randomUUID: () => '12345678-1234-1234-1234-123456789012',
-    })
-
-    const mod = await freshImport()
-    await mod.setAnalyticsUserId('test-user-fnv')
-    // Should not throw — FNV fallback is used
-
-    vi.stubGlobal('crypto', originalCrypto)
-  })
-
-  it('uses FNV fallback when crypto is entirely undefined', async () => {
-    const originalCrypto = globalThis.crypto
-    vi.stubGlobal('crypto', undefined)
-
-    const mod = await freshImport()
-    // Need to mock crypto.randomUUID for getOrCreateAnonymousId
-    // Since crypto is undefined, the demo-user path will fail on randomUUID
-    // Test with a real user ID instead (skips getOrCreateAnonymousId)
-    await mod.setAnalyticsUserId('real-user-no-crypto')
-
-    vi.stubGlobal('crypto', originalCrypto)
-  })
-
-  it('assigns anonymous ID for demo-user', async () => {
-    const mod = await freshImport()
-    await mod.setAnalyticsUserId('demo-user')
-    const anonId = localStorage.getItem('kc-anonymous-user-id')
-    expect(anonId).toBeTruthy()
-  })
-
-  it('assigns anonymous ID for empty string', async () => {
-    const mod = await freshImport()
-    await mod.setAnalyticsUserId('')
-    const anonId = localStorage.getItem('kc-anonymous-user-id')
-    expect(anonId).toBeTruthy()
-  })
-
-  it('propagates user_id to gtag when available', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Make gtag available
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const gtagSpy = vi.fn()
-    window.gtag = gtagSpy
-
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-    gtagSpy.mockClear()
-
-    await mod.setAnalyticsUserId('gtag-user')
-    // Should call gtag('config', ..., { user_id: ... })
-    const configCalls = gtagSpy.mock.calls.filter(([type]) => type === 'config')
-    expect(configCalls.length).toBeGreaterThanOrEqual(1)
-    expect(configCalls[0][2]).toHaveProperty('user_id')
-  })
-})
-
-// ============================================================================
-// tryChunkReloadRecovery — chunk error detection, throttling, recovery failure
-// ============================================================================
-
-describe('tryChunkReloadRecovery via global error handler', () => {
-  it('reloads page on chunk load error', async () => {
-    const reloadSpy = vi.fn()
-    Object.defineProperty(window, 'location', {
-      value: { ...window.location, reload: reloadSpy, pathname: '/test' },
-      writable: true,
-      configurable: true,
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    // Dispatch unhandledrejection with chunk load error
-    const event = new Event('unhandledrejection') as Event & {
-      reason: { message: string }
-    }
-    Object.defineProperty(event, 'reason', {
-      value: { message: 'Failed to fetch dynamically imported module /chunk-abc.js' },
-    })
-    window.dispatchEvent(event)
-
-    expect(reloadSpy).toHaveBeenCalled()
-  })
-
-  it('does not reload when recently reloaded (throttle path exercised)', async () => {
-    // This test exercises the throttle branch in tryChunkReloadRecovery:
-    // when a recent reload timestamp exists and hasn't expired, it skips reload
-    // and emits recovery_failed instead.
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    // Set recent reload timestamp AFTER init to simulate throttle scenario
-    sessionStorage.setItem('ksc-chunk-reload-ts', String(Date.now() - 1000))
-    beaconSpy.mockClear()
-
-    const event = new Event('unhandledrejection') as Event & {
-      reason: { message: string }
-    }
-    Object.defineProperty(event, 'reason', {
-      value: { message: 'Failed to fetch dynamically imported module /chunk-abc.js' },
-    })
-    window.dispatchEvent(event)
-
-    // The throttle branch clears the marker and emits recovery_failed
-    expect(sessionStorage.getItem('ksc-chunk-reload-ts')).toBeNull()
-    // Should have emitted recovery_failed event
-    const recoveryFailed = beaconSpy.mock.calls.some(([url]) => {
-      try {
-        const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-        return decoded.includes('recovery_result') && decoded.includes('failed')
-      } catch { return false }
-    })
-    expect(recoveryFailed).toBe(true)
-  })
-
-  it('emits chunk_load error via runtime handler for Safari messages', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const reloadSpy = vi.fn()
-    Object.defineProperty(window, 'location', {
-      value: { ...window.location, reload: reloadSpy, pathname: '/test' },
-      writable: true,
-      configurable: true,
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    // Fire as a window 'error' event
-    const errorEvent = new ErrorEvent('error', {
-      message: 'Importing a module script failed',
-    })
-    window.dispatchEvent(errorEvent)
-
-    expect(reloadSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// startGlobalErrorTracking — error filtering
-// ============================================================================
-
-describe('global error tracking filters', () => {
-  async function setupErrorTracking() {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-    return { mod, beaconSpy }
-  }
-
-  function dispatchRejection(msg: string, name?: string) {
-    const event = new Event('unhandledrejection') as Event & {
-      reason: { message: string; name?: string }
-    }
-    Object.defineProperty(event, 'reason', {
-      value: { message: msg, ...(name ? { name } : {}) },
-    })
-    window.dispatchEvent(event)
-  }
-
-  function dispatchError(msg: string) {
-    const event = new ErrorEvent('error', { message: msg })
-    window.dispatchEvent(event)
-  }
-
-  it('skips clipboard errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Failed to execute writeText on Clipboard')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBe(0)
-  })
-
-  it('skips AbortError by name', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('The user aborted a request.', 'AbortError')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBe(0)
-  })
-
-  it('skips TimeoutError by name', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('The operation timed out', 'TimeoutError')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBe(0)
-  })
-
-  it('skips Fetch is aborted messages', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Fetch is aborted')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips signal timed out messages', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('signal timed out')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips Load failed messages', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Load failed')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips WebKit URL pattern match errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('The string did not match the expected pattern.')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips JSON parse errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('JSON.parse: unexpected character at line 1')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips "is not valid JSON" errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Unexpected end of input is not valid JSON')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips "JSON Parse error" errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('JSON Parse error: Unexpected token <')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips "Unexpected token" errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Unexpected token < in JSON at position 0')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips ServiceWorker notification errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Failed to execute showNotification on ServiceWorkerRegistration')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips "No active registration" errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('No active registration for this origin')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips WebSocket send-before-connect errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('send was called before connect')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips InvalidStateError errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('InvalidStateError: WebSocket state changed')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips BackendUnavailableError on Netlify', async () => {
-    mockIsNetlifyDeployment = true
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Backend API is currently unavailable')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('does NOT skip BackendUnavailableError on non-Netlify', async () => {
-    mockIsNetlifyDeployment = false
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Backend API is currently unavailable')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('reports genuine unhandled rejections', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchRejection('Cannot read property of undefined')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('skips "Script error." from cross-origin scripts', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchError('Script error.')
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips empty message error events', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    const event = new ErrorEvent('error', { message: '' })
-    window.dispatchEvent(event)
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('reports genuine runtime errors', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchError('TypeError: Cannot read properties of null')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('skips errors already reported by error boundary (dedup)', async () => {
-    const { mod } = await setupErrorTracking()
-    // markErrorReported stores the error in the dedup map — this exercises
-    // the wasAlreadyReported() check in the unhandledrejection handler.
-    // Due to accumulated handlers from prior freshImport() calls in the test suite,
-    // we can't assert exact beacon counts. Instead, verify the code path is exercised
-    // without throwing.
-    mod.markErrorReported('Duplicate error from boundary')
-    expect(() => dispatchRejection('Duplicate error from boundary')).not.toThrow()
-  })
-
-  it('skips clipboard errors in window error handler', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    dispatchError('Failed to execute writeText on Clipboard')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBe(0)
-  })
-
-  it('handles rejection with no reason gracefully', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    const event = new Event('unhandledrejection')
-    Object.defineProperty(event, 'reason', { value: null })
-    window.dispatchEvent(event)
-    // Should handle gracefully (stringifies to 'unknown' or 'null')
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    // May or may not emit depending on the "unknown" fallback
-    expect(true).toBe(true) // just verifying no crash
-  })
-
-  it('handles rejection with string reason', async () => {
-    const { beaconSpy } = await setupErrorTracking()
-    const event = new Event('unhandledrejection')
-    Object.defineProperty(event, 'reason', { value: 'plain string error' })
-    window.dispatchEvent(event)
-    const errorBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_error')
-    })
-    expect(errorBeacons.length).toBeGreaterThanOrEqual(1)
-  })
-})
-
-// ============================================================================
-// checkChunkReloadRecovery — sessionStorage marker
-// ============================================================================
-
-describe('checkChunkReloadRecovery', () => {
-  it('detects recovery marker and stores pending event', async () => {
-    const reloadTime = Date.now() - 300
-    sessionStorage.setItem('ksc-chunk-reload-ts', String(reloadTime))
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-
-    // Marker should be cleared
-    expect(sessionStorage.getItem('ksc-chunk-reload-ts')).toBeNull()
-  })
-
-  it('does nothing when no marker exists', async () => {
-    const mod = await freshImport()
-    // Should not throw
-    expect(() => mod.initAnalytics()).not.toThrow()
-  })
-})
-
-// ============================================================================
-// captureUtmParams — URL param extraction, sessionStorage fallback
-// ============================================================================
-
-describe('captureUtmParams deep', () => {
-  it('captures UTM params from URL and stores in sessionStorage', async () => {
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        search: '?utm_source=test&utm_medium=email',
-        href: 'http://localhost/?utm_source=test&utm_medium=email',
-        hostname: 'localhost',
-        pathname: '/',
-        origin: 'http://localhost',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    const mod = await freshImport()
-    mod.captureUtmParams()
-
-    const stored = sessionStorage.getItem('_ksc_utm')
-    expect(stored).toBeTruthy()
-    const parsed = JSON.parse(stored!)
-    expect(parsed.utm_source).toBe('test')
-    expect(parsed.utm_medium).toBe('email')
-  })
-
-  it('recovers UTM params from sessionStorage on subsequent page loads', async () => {
-    // Ensure URL has NO UTM params so the sessionStorage fallback path is taken
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        search: '',
-        href: 'http://localhost/',
-        hostname: 'localhost',
-        pathname: '/',
-        origin: 'http://localhost',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    sessionStorage.setItem(
-      '_ksc_utm',
-      JSON.stringify({ utm_source: 'cached', utm_campaign: 'test' }),
-    )
-
-    const mod = await freshImport()
-    mod.captureUtmParams()
-    const params = mod.getUtmParams()
-    expect(params.utm_source).toBe('cached')
-    expect(params.utm_campaign).toBe('test')
-  })
-
-  it('truncates UTM values to 100 chars', async () => {
-    const longValue = 'x'.repeat(200)
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        search: `?utm_source=${longValue}`,
-        href: `http://localhost/?utm_source=${longValue}`,
-        hostname: 'localhost',
-        pathname: '/',
-        origin: 'http://localhost',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    const mod = await freshImport()
-    mod.captureUtmParams()
-    const params = mod.getUtmParams()
-    expect(params.utm_source?.length).toBeLessThanOrEqual(100)
-  })
-
-  it('captures all 5 UTM parameters', async () => {
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        search:
-          '?utm_source=src&utm_medium=med&utm_campaign=camp&utm_term=trm&utm_content=cnt',
-        href: 'http://localhost/?utm_source=src&utm_medium=med&utm_campaign=camp&utm_term=trm&utm_content=cnt',
-        hostname: 'localhost',
-        pathname: '/',
-        origin: 'http://localhost',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    const mod = await freshImport()
-    mod.captureUtmParams()
-    const params = mod.getUtmParams()
-    expect(params.utm_source).toBe('src')
-    expect(params.utm_medium).toBe('med')
-    expect(params.utm_campaign).toBe('camp')
-    expect(params.utm_term).toBe('trm')
-    expect(params.utm_content).toBe('cnt')
-  })
-})
-
-// ============================================================================
-// setAnalyticsOptOut — deep: cookie cleanup, engagement stop
-// ============================================================================
-
-describe('setAnalyticsOptOut deep', () => {
-  it('clears _ga and _ksc cookies on opt-out', async () => {
-    // Set some cookies
-    document.cookie = '_ga=GA1.1.12345;path=/'
-    document.cookie = '_ksc_cid=test-cid;path=/'
-    document.cookie = 'unrelated=keep;path=/'
-
-    const mod = await freshImport()
-    mod.setAnalyticsOptOut(true)
-
-    // Verify opt-out flag is set
-    expect(localStorage.getItem('kc-analytics-opt-out')).toBe('true')
-  })
-
-  it('dispatches settings-changed event', async () => {
-    const handler = vi.fn()
-    window.addEventListener('kubestellar-settings-changed', handler)
-
-    const mod = await freshImport()
-    mod.setAnalyticsOptOut(true)
-    expect(handler).toHaveBeenCalledTimes(1)
-
-    mod.setAnalyticsOptOut(false)
-    expect(handler).toHaveBeenCalledTimes(2)
-
-    window.removeEventListener('kubestellar-settings-changed', handler)
-  })
-})
-
-// ============================================================================
-// updateAnalyticsIds
-// ============================================================================
-
-describe('updateAnalyticsIds deep', () => {
-  it('updates ga4 measurement ID (does not throw)', async () => {
-    const mod = await freshImport()
-    // updateAnalyticsIds is called by BrandingProvider before init in production.
-    // Since module-level state may not fully reset with vi.mock, just verify the call works.
-    expect(() => mod.updateAnalyticsIds({ ga4MeasurementId: 'G-CUSTOM123' })).not.toThrow()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    // Verify gtag script was appended (regardless of which measurement ID)
-    expect(window.dataLayer).toBeDefined()
-  })
-
-  it('updates umami website ID', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.updateAnalyticsIds({ umamiWebsiteId: 'custom-umami-id' })
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    const umamiScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/ksc'),
-      ) as HTMLScriptElement | undefined
-
-    expect(umamiScript?.dataset.websiteId).toBe('custom-umami-id')
-  })
-
-  it('does not override with empty strings', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.updateAnalyticsIds({ ga4MeasurementId: '', umamiWebsiteId: '' })
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // Should still use defaults
-    const gtagScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    // Default is G-PXWNVQ8D1T
-    expect(gtagScript?.src).toContain('G-PXWNVQ8D1T')
-  })
-})
-
-// ============================================================================
-// setAnalyticsUserProperties — gtag propagation
-// ============================================================================
-
-describe('setAnalyticsUserProperties gtag propagation', () => {
-  it('propagates to gtag when available', async () => {
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const gtagSpy = vi.fn()
-    window.gtag = gtagSpy
-
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-    gtagSpy.mockClear()
-
-    mod.setAnalyticsUserProperties({ role: 'admin' })
-    const setCalls = gtagSpy.mock.calls.filter(([type]) => type === 'set')
-    expect(setCalls.length).toBeGreaterThanOrEqual(1)
-    expect(setCalls[0][1]).toBe('user_properties')
-    expect(setCalls[0][2]).toEqual({ role: 'admin' })
-  })
-})
-
-// ============================================================================
-// emitDemoModeToggled — updates userProperties
-// ============================================================================
-
-describe('emitDemoModeToggled updates user properties', () => {
-  it('fires event and updates internal demo_mode property', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitDemoModeToggled(true)
-    expect(beaconSpy).toHaveBeenCalled()
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('en=ksc_demo_mode_toggled')
-  })
-})
-
-// ============================================================================
-// emitSessionContext — deduplication via sessionStorage
-// ============================================================================
-
-describe('emitSessionContext dedup', () => {
-  it('sends ksc_session_start only once per tab', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitSessionContext('binary', 'stable')
-    const firstCallCount = beaconSpy.mock.calls.length
-
-    mod.emitSessionContext('binary', 'stable')
-    const secondCallCount = beaconSpy.mock.calls.length
-
-    // Second call should not emit session_start (only sets user properties)
-    // The difference should be smaller since session_start is deduped
-    expect(secondCallCount - firstCallCount).toBeLessThan(firstCallCount)
-  })
-})
-
-// ============================================================================
-// emitDeveloperSession — guards
-// ============================================================================
-
-describe('emitDeveloperSession guards', () => {
-  it('skips when already sent (localStorage dedup)', async () => {
-    localStorage.setItem('ksc-dev-session-sent', '1')
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitDeveloperSession()
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-
-  it('skips when not on localhost', async () => {
-    Object.defineProperty(window, 'location', {
-      value: {
-        ...window.location,
-        hostname: 'console.kubestellar.io',
-        href: 'https://console.kubestellar.io/',
-        pathname: '/',
-        origin: 'https://console.kubestellar.io',
-        search: '',
-      },
-      writable: true,
-      configurable: true,
-    })
-
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitDeveloperSession()
-    // Should not emit ksc_developer_session for non-localhost
-    const devSessionBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_developer_session')
-    })
-    expect(devSessionBeacons.length).toBe(0)
-  })
-
-  it('skips on localhost when in forced demo mode without token', async () => {
-    mockIsDemoMode = true
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitDeveloperSession()
-    const devSessionBeacons = beaconSpy.mock.calls.filter(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=ksc_developer_session')
-    })
-    expect(devSessionBeacons.length).toBe(0)
-  })
-})
-
-// ============================================================================
-// emitAgentProvidersDetected — capability bitmask
-// ============================================================================
-
-describe('emitAgentProvidersDetected deep', () => {
-  it('categorizes CLI vs API providers by capability bitmask', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitAgentProvidersDetected([
-      { name: 'claude-code', displayName: 'Claude Code', capabilities: 3 }, // CHAT + TOOL_EXEC
-      { name: 'openai', displayName: 'OpenAI', capabilities: 1 }, // CHAT only
-      { name: 'cursor', displayName: 'Cursor', capabilities: 2 }, // TOOL_EXEC only
-    ])
-
-    expect(beaconSpy).toHaveBeenCalled()
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('en=ksc_agent_providers_detected')
-    expect(decoded).toContain('epn.provider_count=3')
-    // CLI = tool_exec capable: claude-code, cursor
-    expect(decoded).toContain('ep.cli_providers=claude-code%2Ccursor')
-    // API = chat-only: openai
-    expect(decoded).toContain('ep.api_providers=openai')
-  })
-
-  it('returns "none" when no CLI or API providers', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitAgentProvidersDetected([
-      { name: 'unknown', displayName: 'Unknown', capabilities: 0 },
-    ])
-
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('ep.cli_providers=none')
-    expect(decoded).toContain('ep.api_providers=none')
-  })
-
-  it('early-returns for null/empty providers', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitAgentProvidersDetected([])
-    expect(beaconSpy).not.toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// emitNPSSurveyShown / emitNPSResponse / emitNPSDismissed
-// ============================================================================
-
-describe('NPS survey events', () => {
-  async function setupProxy() {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-    return { mod, beaconSpy }
-  }
-
-  it('emitNPSSurveyShown sends event', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitNPSSurveyShown()
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('emitNPSResponse sends score and category', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitNPSResponse(9, 'promoter', 42)
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('en=ksc_nps_response')
-    expect(decoded).toContain('epn.nps_score=9')
-    expect(decoded).toContain('ep.nps_category=promoter')
-    expect(decoded).toContain('epn.nps_feedback_length=42')
-  })
-
-  it('emitNPSResponse works without feedbackLength', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitNPSResponse(5, 'passive')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('emitNPSDismissed sends dismiss count', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitNPSDismissed(3)
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('en=ksc_nps_dismissed')
-    expect(decoded).toContain('epn.dismiss_count=3')
-  })
-})
-
-// ============================================================================
-// Orbit / GroundControl events
-// ============================================================================
-
-describe('Orbit and GroundControl events', () => {
-  async function setupProxy() {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-    return { mod, beaconSpy }
-  }
-
-  it('emitOrbitMissionCreated sends event', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitOrbitMissionCreated('security-scan', 'daily')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('emitOrbitMissionRun sends event', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitOrbitMissionRun('security-scan', 'success')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('emitGroundControlDashboardCreated sends event', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitGroundControlDashboardCreated(5)
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-
-  it('emitGroundControlCardRequestOpened sends event', async () => {
-    const { mod, beaconSpy } = await setupProxy()
-    mod.emitGroundControlCardRequestOpened('kubestellar')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// emitBlogPostClicked
-// ============================================================================
-
-describe('emitBlogPostClicked', () => {
-  it('sends event with blog title', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-    beaconSpy.mockClear()
-
-    mod.emitBlogPostClicked('My Test Blog Post')
-    const url = beaconSpy.mock.calls[0][0] as string
-    const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-    expect(decoded).toContain('en=ksc_blog_post_clicked')
-    expect(decoded).toContain('blog_title')
-  })
-})
-
-// ============================================================================
-// Engaged session threshold via proxy
-// ============================================================================
-
-describe('engaged session tracking', () => {
-  it('sets seg=1 after 10s of engagement', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100) // force proxy mode
-    beaconSpy.mockClear()
-
-    // Simulate 11 seconds of activity to cross engaged threshold
-    vi.advanceTimersByTime(11000)
-    // Keep "active" by triggering interactions
-    document.dispatchEvent(new Event('mousedown'))
-
-    mod.emitPageView('/engaged')
-    if (beaconSpy.mock.calls.length > 0) {
-      const url = beaconSpy.mock.calls[beaconSpy.mock.calls.length - 1][0] as string
-      const decoded = atob(decodeURIComponent(url.split('d=')[1]))
-      // After 10s of engagement, seg=1 should appear
-      expect(decoded).toContain('seg=1')
-    }
-  })
-})
-
-// ============================================================================
-// user_engagement event — getAndResetEngagementMs vs peekEngagementMs
-// ============================================================================
-
-describe('engagement time in events', () => {
-  it('uses getAndResetEngagementMs for user_engagement events in proxy', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100) // proxy mode
-
-    // Accumulate some engagement time
-    vi.advanceTimersByTime(3000)
-    document.dispatchEvent(new Event('mousedown')) // keep active
-    beaconSpy.mockClear()
-
-    // Trigger tab hidden which calls emitUserEngagement
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'hidden',
-      configurable: true,
-    })
-    document.dispatchEvent(new Event('visibilitychange'))
-
-    // Check for user_engagement event with _et parameter
-    const engagementBeacon = beaconSpy.mock.calls.find(([url]) => {
-      const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-      return decoded.includes('en=user_engagement')
-    })
-
-    if (engagementBeacon) {
-      const decoded = atob(
-        decodeURIComponent((engagementBeacon[0] as string).split('d=')[1]),
-      )
-      expect(decoded).toContain('_et=')
-    }
-
-    Object.defineProperty(document, 'visibilityState', {
-      value: 'visible',
-      configurable: true,
-    })
-  })
-})
-
-// ============================================================================
-// markGtagDecided idempotency
-// ============================================================================
-
-describe('markGtagDecided idempotency', () => {
-  it('only the first call takes effect', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const appendSpy = vi.spyOn(document.head, 'appendChild')
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-
-    // First decision: timeout triggers proxy mode
-    vi.advanceTimersByTime(5100)
-
-    // Now if gtag script loads after timeout, it should NOT override
-    ;(window as Record<string, unknown>).google_tag_manager = {}
-    const firstScript = appendSpy.mock.calls
-      .map(([el]) => el)
-      .find(
-        (el) =>
-          el instanceof HTMLScriptElement && el.src.includes('/api/gtag'),
-      ) as HTMLScriptElement | undefined
-
-    if (firstScript?.onload) {
-      ;(firstScript.onload as () => void)(new Event('load'))
-    }
-    vi.advanceTimersByTime(150)
-
-    // Events should still go via proxy (first decision was false)
-    beaconSpy.mockClear()
-    mod.emitCardAdded('test', 'manual')
-    expect(beaconSpy).toHaveBeenCalled()
-  })
-})
-
-// ============================================================================
-// wasAlreadyReported — expiry after DEDUP_EXPIRY_MS (5s)
-// ============================================================================
-
-describe('wasAlreadyReported dedup expiry', () => {
-  it('reports error after dedup window expires', async () => {
-    const beaconSpy = vi.fn(() => true)
-    vi.stubGlobal('navigator', {
-      ...navigator,
-      sendBeacon: beaconSpy,
-      language: 'en-US',
-    })
-
-    const mod = await freshImport()
-    mod.initAnalytics()
-    document.dispatchEvent(new Event('mousedown'))
-    vi.advanceTimersByTime(5100)
-
-    // Mark error as reported
-    mod.markErrorReported('dedup-unique-xyz-test')
-    const callsAfterMark = beaconSpy.mock.calls.length
-
-    // Within 5s window — should be skipped by THIS module's handler
-    const event1 = new Event('unhandledrejection')
-    Object.defineProperty(event1, 'reason', {
-      value: { message: 'dedup-unique-xyz-test' },
-    })
-    window.dispatchEvent(event1)
-    // Count NEW calls only (other handlers from prior freshImport may fire)
-    const newCalls1 = beaconSpy.mock.calls.slice(callsAfterMark)
-    const dedupedErrors1 = newCalls1.filter(([url]) => {
-      try {
-        const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-        return decoded.includes('unhandled_rejection') && decoded.includes('dedup-unique-xyz-test')
-      } catch { return false }
-    })
-    // The current module's handler should have skipped it (deduped)
-    // Note: older handlers from prior tests may or may not emit it
-    // The key behavior is that the dedup mechanism works within a single module instance
-
-    // Advance past dedup window (5s)
-    vi.advanceTimersByTime(6000)
-    const callsAfterExpiry = beaconSpy.mock.calls.length
-
-    const event2 = new Event('unhandledrejection')
-    Object.defineProperty(event2, 'reason', {
-      value: { message: 'dedup-unique-xyz-test' },
-    })
-    window.dispatchEvent(event2)
-    const newCalls2 = beaconSpy.mock.calls.slice(callsAfterExpiry)
-    const reportedErrors2 = newCalls2.filter(([url]) => {
-      try {
-        const decoded = atob(decodeURIComponent((url as string).split('d=')[1]))
-        return decoded.includes('unhandled_rejection')
-      } catch { return false }
-    })
-    // After expiry, at least some handler should report it
-    expect(reportedErrors2.length).toBeGreaterThanOrEqual(1)
-  })
-})
